requirements.txt
azure-functions==1.23.0
load-dotenv==0.1.0
MarkupSafe==3.0.2
python-dotenv==1.1.1
Werkzeug==3.1.3
